Add files via upload
toast.py is the initial testing of Windows toast notifications with the win10toast library, and will display 10 numbered toast notifications in sequence.
durationBypassToast.py was made to test if the duration parameter for toast notifications would overpower the notification duration as dictated by the corresponding registry, which was updated via the Registry Editor. The findings were included as comments in that file.